Add scan of the ROC curve

+++ utils/scan_parameters.py
@@ -8,6 +8,7 @@
 #
 
 from __future__ import print_function
+from sklearn.metrics import roc_curve, auc
 import pprint
 import pandas
 import json
@@ -18,6 +19,9 @@ import collections
 import itertools
 import subprocess
 import xml.etree.ElementTree as ET
+import matplotlib
+matplotlib.use('agg')
+import matplotlib.pyplot as plt
 
 ToRun = True
 
@@ -87,6 +91,12 @@ def ExtractFromXML(df, key, report_file):
         if k not in df:
             df[k] = None
 
+    ## Check file exits
+    if not os.path.isfile(report_file):
+        for k, v in xmlpath.items():
+            df.at[key, k] = None
+        return False
+
     ## Parsing XML file
     tree = ET.parse(report_file)
     root = tree.getroot()
@@ -96,6 +106,32 @@ def ExtractFromXML(df, key, report_file):
             path = path.find(i)
         df.at[key, k] = path.text
 
+def ExtractROC(df, key, output_filename):
+    truth_filename = "../data_to_txt_2NN/KERAS_check_best_model_truth_labels.dat"
+    predict_filename = "../data_to_txt_2NN/KERAS_check_best_model_predictions.dat"
+
+    if "AUC" not in df:
+        df["AUC"] = None
+
+    ## Check file exits
+    if not os.path.isfile(output_filename):
+        df.at[key, "AUC"] = None
+        return False
+
+    truth_df = pandas.read_csv(truth_filename, header=None)
+    predict_df = pandas.read_csv(predict_filename, header=None)
+    output_df = pandas.read_csv(output_filename, header=None)
+
+    ## Expected AUC from keras
+    # fpr, tpr, threshold = roc_curve(truth_df,predict_df)
+
+    dfpr, dtpr, dthreshold = roc_curve(truth_df,output_df.iloc[:, 0])
+    dauc = auc(dfpr, dtpr)
+    df.at[key, "AUC"] = dauc
+
+    plt.plot(dtpr,dfpr,label='%s point, auc = %.1f%%'%( key, dauc * 100))
+    plt.savefig("%s.pdf" % key)
+
 
 def RunProjs(projs):
     ##
@@ -104,7 +140,10 @@ def RunProjs(projs):
     for k, v in projs.items():
         PrepareYaml(k, v)
         pwd = os.getcwd()
-        if ToRun:
+        report_filename = "%s/../../keras-to-hls/%s/myproject_prj/solution1/syn/report/myproject_csynth.xml" % (pwd, k)
+        output_filename = "%s/../../keras-to-hls/%s/myproject_prj/solution1/csim/build/res.dat" % (pwd,k)
+
+        if ToRun and not os.path.exists(report_filename):
             ymltorun = "%s/%s.yml" % (pwd, k)
             outlog = open("%s.stdout" % k, 'w')
             errlog = open("%s.stderr" % k, 'w')
@@ -120,9 +159,17 @@ def RunProjs(projs):
                 subprocess.call("rm -rf %s" % autopilot,
                             stdout = outlog, stderr=errlog, shell=True)
 
+            ## Remove large data input file, whose size scale up with precision 
+            wdbfile = "%s/../../keras-to-hls/%s/myproject_prj/solution1/sim/verilog/myproject.wdb" % (pwd, k)
+            if os.path.exists(autopilot):
+                subprocess.call("rm -rf %s" % autopilot,
+                            stdout = outlog, stderr=errlog, shell=True)
 
-        report_filename = "%s/../../keras-to-hls/%s/myproject_prj/solution1/syn/report/myproject_csynth.xml" % (pwd, k)
+        # Extract XML file
         ExtractFromXML(df, k,  report_filename)
+        # Extract ROC curve
+        ExtractROC(df, k, output_filename)
+
     return df
 
 if __name__ == "__main__":
@@ -139,4 +186,3 @@ if __name__ == "__main__":
     projs = FormVariation(config)
     df = RunProjs(projs)
     df.to_csv("output.csv")
-
